AO3-5905 Move WorkStats methods into Work.
